Compiler warning fixes

+++ PlayerMovement.h
@@ -13,9 +13,9 @@ public:
 
 	void LoadMovementTextures();
 		
-	int PlayerMovement::walk(int x, int y, int gd, int gy, int screenWidth, int screenHeight);
+	double PlayerMovement::walk(int x, int y, int gd, int gy, int screenWidth, int screenHeight);
 
-	int PlayerMovement::walky(int x, int y, int gd, int gy, int screenWidth, int screenHeight);
+	double PlayerMovement::walky(int x, int y, int gd, int gy, int screenWidth, int screenHeight);
 
 	int doXWalkRight(int gd, int screenWidth);
